mds: Merge parent snaps into children on new parent snap or when opening child
This is a major shift: previously, we had to keep our current and all past
parents in-memory whenever doing file IO, because we relied on them to build
up accurate lists of snapids for RADOS. Now, we push those snapids into
lower layers at every opportunity (either when they're already in-memory as
snapshots are created, or when we open them). That means we won't need
to go look up past parents any more!

The tradeoff is that by duplicating snapids in more metadata, we're increasing
disk/memory usage a little bit. Tough luck.

Signed-off-by: Greg Farnum <gfarnum@redhat.com>