TODO: Revert "mds: update snaptable when renaming snapshot"
This reverts commit 47c2cecc21e0546a53e327e4adfb9f0f38f3cae9.

Updating snapshot names doesn't work properly since the SnapInfo is duplicated
across many inodes now. It also makes the SnapServer a little more complicated
to follow. Kill it.

TODO: This didn't change the struct_v level encoding, but we are removing
the possibility of handling a TABLE_OP_UPDATE correctly and that's
obviously a caompatibility issue. Also, being allowed to do renames is nice!

Conflicts:
	src/mds/Server.cc