ceph_mon: "mon host" config should be honored
When no mon section are declared, the current behavior was to bind on the public network.
The "mon host" configuration should be honored in such case.
This patch is about reading the "mon host" configuration option before using the public network as fall back.

In such case, a typical output looks like :

	2016-11-25 17:16:17.869587 7f87502c5580  0 mon.ceph-mon0 does not exist in monmap, will attempt to join an existing cluster
	2016-11-25 17:16:17.869621 7f87502c5580  0 using mon host 172.16.0.10:0/0 -> 172.16.0.10:6789/0
	2016-11-25 17:16:17.869723 7f87502c5580  0 starting mon.ceph-mon0 rank -1 at 172.16.0.10:6789/0 mon_data /var/lib/ceph/mon/ceph-ceph-mon0 fsid 8a5bd9e5-001a-431a-9f84-36fef7fc7bc4

fixes: #18043

Tested-by: Sebastien Han <seb@redhat.com>
Signed-off-by: Erwan Velu <erwan@redhat.com>